+++ src/node.h
@@ -8,20 +8,20 @@ private:
     int phase;
     int job;
     int day;
-    unsigned int start;
-    unsigned int end;
+    long int start;
+    long int end;
     void destroy(Node *node);
 public:
     /* For leaf nodes */
-    Node(int phase, int job, int day, unsigned int start, unsigned int end);
+    Node(int phase, int job, int day, long int start, long int end);
     
     /* Called by root */
     void destroy();
     int getPhase();
     int getJob();
     int getDay();
-    unsigned int getStart();
-    unsigned int getEnd();
+    long int getStart();
+    long int getEnd();
 
 
 };
